Music-Rework-Update-1.5
updates music based on player feedback and desires. all music is now tuned to -10 DB, to reduce the volume., future music updates will be tuned on exact same pregorative.
Music now fits more in town to the dark theme of the town, and is also more ambient.
combat music remains same, but quieter. Other town music and general single music file is replaced with fine score from ATRIUM CARCERI.

In next update. 1.6, i will talk to community extensively about the castle music, and see if we want a change. If there isnt any problem with castle OST and nothing is desired to be changed, wilderness music is going to be discussed and applied for 1.6.