Underworld Changes
Adds more variety to spirit names so that when people ahelp "Dejected is [doing bad thing] there won't be 20 spirits with that exact name.

Removes the ability for spirits to change movement speed or middle mouse intents.

Drastically reduces the number of tolls that need to be destroyed before triggering a respawn. Tolls will respawn when they hit 7 left.