Vault fixes, duplicate item penalties.
If an item worth 0 was placed into the vault, the vault would stop processing early and not pay out.

Adds a penalty for multiple of the same item type in the treasury, encouraging stewards to diversify the items they place inside. Each item of the same type will earn 33% cumulative less worth every treasury tick.

Coins in the treasury will not accrue interest.