Re-implements poisoned arrows, craftable mortar + pestle
You can now craft a mortar and a pestle from small stones, in order to crush things like poisonous berries. The berry poison can then be poured in a bucket and used to make poisoned arrows, or by crafty malcontents to poison the contents of a barrel. While I have written and included the code for poisoning crossbow bolts, it has been commented out for now because devs may wish to tweak crossbows further and they really don't need the added lethality.