ALMOST THERE
the pike for now only stores one head. you can see the head sprite on the pike when put on it now, and you can examine the pike to see the head.
you can also LMB the pike again with an empty hand to take down the head, but the current issue now is the head doesn't end up in your hands...or anywhere.

BUT IT'S ALMOST THERE