Buff arming cap
This is basically the same as coif but can also be worn on the face (cope for armor layering)

It's whatever. It can have the same shit as regular coif. They're the same fucking thing really though arming cap should technically only cover the head/ears/hair and not neck.